+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Add `sign_hashx` and `decorated_signature_from_preimage` to
+   `FeeBumpTransaction` and `TransactionEnvelope`.
 
 
 ## [0.3.0] - 2020-07-18